// web/lib/generated/deployments.ts
// Generated by scripts/export-abis.ts — do not edit.
export const deployments = {
  "968": {
    "chainId": 968,
    "explorer": "https://scan.bohr.life",
    "deployedAtBlock": 20204319,
    "deployedAt": "2026-08-17T17:07:10.874Z",
    "deployer": "0x153ee46725bf53a32d2269200e5e992955e2eb5b",
    "contracts": {
      "IdentityRegistry": "0xe06a0801706679e73dd04917a63aa796788f5ffd",
      "ReputationRegistry": "0x8a518ab8c75c562170d71a88151cf1611b811fbe",
      "ValidationRegistry": "0x84b1bb992c5bb33e4cd05b32cc969683d1f3c511",
      "LetterOfCredit": "0x1145970c4eb218a7d1a05245503a0fc05e7b8e99",
      "ServiceVendor": "0xdd4ffa32cd6e8b2d0afeb1ab277683ad16bb4d61"
    }
  }
} as const;